chore: move beacon handle
